Connecting thermometer related code moved to addThermometerModule
