+++ addThermometer/addThermometerPresentation/src/main/java/com/klewerro/mitemperature2alt/addThermometerPresentation/ConnectThermometerEvent.kt
@@ -1,4 +1,4 @@
-package com.klewerro.mitemperature2alt.presentation.addThermometer
+package com.klewerro.mitemperature2alt.addThermometerPresentation
 
 sealed class ConnectThermometerEvent {
     data object ConnectToDevice : ConnectThermometerEvent()
