+++ addThermometer/addThermometerPresentation/src/main/java/com/klewerro/mitemperature2alt/addThermometerPresentation/ConnectThermometerViewModel.kt
@@ -1,15 +1,15 @@
-package com.klewerro.mitemperature2alt.presentation.addThermometer
+package com.klewerro.mitemperature2alt.addThermometerPresentation
 
 import androidx.lifecycle.SavedStateHandle
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.klewerro.mitemperature2alt.coreUi.R
+import com.klewerro.mitemperature2alt.coreUi.UiConstants
 import com.klewerro.mitemperature2alt.coreUi.util.UiText
 import com.klewerro.mitemperature2alt.domain.usecase.thermometer.connect.ConnectToDeviceUseCase
 import com.klewerro.mitemperature2alt.domain.usecase.thermometer.operations.ReadCurrentThermometerStatusUseCase
 import com.klewerro.mitemperature2alt.domain.usecase.thermometer.persistence.SaveThermometerUseCase
 import com.klewerro.mitemperature2alt.domain.util.DispatcherProvider
-import com.klewerro.mitemperature2alt.presentation.navigation.Route
 import dagger.hilt.android.lifecycle.HiltViewModel
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
@@ -31,7 +31,7 @@ class ConnectThermometerViewModel @Inject constructor(
     private val _state = MutableStateFlow(ConnectThermometerState())
     val state = combine(
         _state,
-        savedState.getStateFlow(Route.ConnectDeviceRoutes.PARAM_ADDRESS, "")
+        savedState.getStateFlow(UiConstants.NAV_PARAM_ADDRESS, "")
     ) { stateValue, address ->
         stateValue.copy(
             thermometerAddress = address
