+++ addThermometer/addThermometerPresentation/src/main/java/com/klewerro/mitemperature2alt/addThermometerPresentation/ThermometerConnectingScreen.kt
@@ -1,4 +1,4 @@
-package com.klewerro.mitemperature2alt.presentation.addThermometer.connecting
+package com.klewerro.mitemperature2alt.addThermometerPresentation
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.AnimatedVisibility
@@ -33,13 +33,8 @@ import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import com.klewerro.mitemperature2alt.R
 import com.klewerro.mitemperature2alt.coreUi.LocalSpacing
 import com.klewerro.mitemperature2alt.coreUi.theme.MiTemperature2AltTheme
-import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectThermometerEvent
-import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectThermometerState
-import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectThermometerViewModel
-import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectingStatus
 import kotlinx.coroutines.delay
 import com.klewerro.mitemperature2alt.coreUi.R as RCore
