+++ addThermometer/addThermometerPresentation/src/main/res/drawable/ic_round_error_24.xml
