+++ addThermometer/addThermometerPresentation/src/main/res/drawable/thermometer_image.webp
