+++ app/src/main/java/com/klewerro/mitemperature2alt/MainActivity.kt
@@ -29,11 +29,12 @@ import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
 import androidx.navigation.compose.rememberNavController
 import androidx.navigation.navArgument
+import com.klewerro.mitemperature2alt.addThermometerPresentation.ConnectThermometerViewModel
+import com.klewerro.mitemperature2alt.addThermometerPresentation.ThermometerConnectingScreen
+import com.klewerro.mitemperature2alt.addThermometerPresentation.name.ConnectThermometerNameScreen
 import com.klewerro.mitemperature2alt.addThermometerPresentation.search.SearchThermometersScreen
+import com.klewerro.mitemperature2alt.coreUi.UiConstants
 import com.klewerro.mitemperature2alt.coreUi.theme.MiTemperature2AltTheme
-import com.klewerro.mitemperature2alt.presentation.addThermometer.ConnectThermometerViewModel
-import com.klewerro.mitemperature2alt.presentation.addThermometer.connecting.ThermometerConnectingScreen
-import com.klewerro.mitemperature2alt.presentation.addThermometer.name.ConnectThermometerNameScreen
 import com.klewerro.mitemperature2alt.presentation.mainscreen.BleOperationsViewModel
 import com.klewerro.mitemperature2alt.presentation.mainscreen.MainScreen
 import com.klewerro.mitemperature2alt.presentation.mainscreen.TopBar
@@ -117,7 +118,7 @@ class MainActivity : ComponentActivity() {
             route = Route.ConnectDeviceRoutes.ConnectDeviceGraph.fullRoute,
             startDestination = Route.ConnectDeviceRoutes.Connecting.fullRoute,
             arguments = listOf(
-                navArgument(Route.ConnectDeviceRoutes.PARAM_ADDRESS) {
+                navArgument(UiConstants.NAV_PARAM_ADDRESS) {
                     type = NavType.StringType
                 }
             )
