+++ app/src/main/java/com/klewerro/mitemperature2alt/presentation/mainscreen/components/MainScreenThermometerBox.kt
@@ -18,9 +18,9 @@ import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
+import com.klewerro.mitemperature2alt.coreUi.components.ThermometerBox
 import com.klewerro.mitemperature2alt.coreUi.theme.MiTemperature2AltTheme
 import com.klewerro.mitemperature2alt.domain.model.ThermometerDevice
-import com.klewerro.mitemperature2alt.presentation.components.ThermometerBox
 
 @Composable
 fun MainScreenThermometerBox(
