+++ app/src/main/java/com/klewerro/mitemperature2alt/presentation/navigation/Route.kt
@@ -3,6 +3,7 @@ package com.klewerro.mitemperature2alt.presentation.navigation
 import androidx.annotation.StringRes
 import androidx.navigation.NavController
 import com.klewerro.mitemperature2alt.coreUi.R
+import com.klewerro.mitemperature2alt.coreUi.UiConstants
 
 sealed class Route(
     @StringRes val screenName: Int,
@@ -32,12 +33,12 @@ sealed class Route(
     }
 
     sealed class ConnectDeviceRoutes {
-        data object ConnectDeviceGraph : Route(R.string.connect, "connect", PARAM_ADDRESS)
+        data object ConnectDeviceGraph : Route(
+            R.string.connect,
+            "connect",
+            UiConstants.NAV_PARAM_ADDRESS
+        )
         data object Connecting : Route(R.string.connecting_to_thermometer, "connect_connecting")
         data object SetName : Route(R.string.set_thermometer_name, "connect_name")
-
-        companion object NavigationParameters {
-            const val PARAM_ADDRESS = "address"
-        }
     }
 }
