+++ coreUi/src/main/java/com/klewerro/mitemperature2alt/coreUi/UiConstants.kt
@@ -1,0 +1,5 @@
+package com.klewerro.mitemperature2alt.coreUi
+
+object UiConstants {
+    const val NAV_PARAM_ADDRESS = "address"
+}
